% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{args_drop_else}
\alias{args_drop_else}
\title{Removes \code{.else} from arguments list}
\usage{
args_drop_else(args)
}
\arguments{
\item{args}{\code{\link{list}}\cr arguments}
}
\value{
last of \code{args} without \code{.else}
}
\description{
#' Removes \code{.else} from arguments list. Technically, last unnamed
list object is removed.
}
