% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_thenelse_OX}
\alias{check_thenelse_OX}
\title{Checks .then and .else for OX}
\usage{
check_thenelse_OX(.then, .else)
}
\description{
Checks if .then and .else are fine. If they have supported classes and
if their length is the same.
}
