% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_else}
\alias{get_else}
\title{Get \code{.else} argument}
\usage{
get_else(args)
}
\arguments{
\item{args}{\code{\link{list}}\cr arguments}
}
\value{
last unnamed \code{args} list element
}
\description{
Gets \code{.else} argument from the arguments list. Technically, last unnamed
list object is returned.
}
