% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{ie}
\alias{ie}
\alias{ei}
\title{Short hand \code{ifelse}}
\usage{
ie(fun, ...)

ei(fun, ...)
}
\arguments{
\item{fun}{\code{\link{function}}\cr}

\item{...\cr}{\itemize{
\item object from the first argument will be returned if evaluation of \code{fun}
returns FALSE. This value is passed to the \code{fun} as a first argument.
NOTE: If it is not a first argument of the \code{fun} please name it so that
\code{ie} can pass it in a right way.
\item object from the second argument will be returned if function returns TRUE.
\item other arguments are passed to the \code{fun} in the same order. NOTE: if order
of arguments is different, please name them accordingly.
}}
}
\description{
Short hand \code{ifelse}
}
\examples{
# return "y" if "x" is a character
ie(is, "x", "y", "character")

has_nchars <- function(x, min, max) nchar(x) >= min && nchar(x) <= max
ie(has_nchars, "test", "test passed", min = 3, max = 5)
}
