% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{invert_indices}
\alias{invert_indices}
\title{Indices inversion}
\usage{
invert_indices(idx)
}
\arguments{
\item{idx}{(\code{logical}, \code{integer})}
}
\value{
same class as \code{idx}
}
\description{
Inverts the indices to select the rest. For example when \code{idx} is \code{logical}
then \verb{[!idx]}, if \code{integer} then \verb{[-idx]}
}
\keyword{internal}
