% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox.R
\name{ox}
\alias{ox}
\title{Creates an ox object from a parsed OpenClinica xml file}
\usage{
ox(parsed_xml)
}
\arguments{
\item{parsed_xml}{}
}
\value{
An ox object, which is a list of two elements: data and metadata. The data element is a
dataframe containig all clinical data. The metadata element is a list of
dataframes (and a vector with global variables), documenting events, forms,
groups, items, codelists, measurement units and sites.
}
\description{
Returns and ox object from a parsed OpenClinica xml file
provided as argument
}
\examples{

}
