% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_alias_ns.R
\name{ox_alias_default_ns}
\alias{ox_alias_default_ns}
\title{Title}
\usage{
ox_alias_default_ns(parsed_xml)
}
\arguments{
\item{parsed_xml}{an object of class XMLInternalDocument}
}
\value{
default namespace alias
}
\description{
Title
}
\examples{

}
