% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_audit_log.R
\name{ox_audit_log}
\alias{ox_audit_log}
\title{Audit log in a dataframe}
\usage{
ox_audit_log(parsed_xml)
}
\arguments{
\item{parsed_xml}{An object of class \code{XMLInternalDocument}, as returned
by \code{XML::xmlParse()}.}
}
\value{
A dataframe.
}
\description{
Returns a dataframe with study audit log from a parsed OpenClinica
odm1.3_full .xml export file.
}
\examples{
# The example odm1.3_full xml file address
my_file <- system.file("extdata",
                       "odm1.3_clinical_ext_example.xml",
                       package = "ox",
                       mustWork = TRUE)

# Parsing the xml file
library(XML)
doc <- xmlParse(my_file)

# Audit log in a dataframe
audit_log <- ox_audit_log(doc)
head(audit_log)
}
