% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_codelist.R
\name{ox_codelist}
\alias{ox_codelist}
\title{Title}
\usage{
ox_codelist(parsed_xml)
}
\arguments{
\item{parsed_xml}{an object of class XMLInternalDocument}
}
\value{
dataframe
}
\description{
Title
}
\examples{

}
