% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_codelist_item.R
\name{ox_codelist_item}
\alias{ox_codelist_item}
\title{Title}
\usage{
ox_codelist_item(parsed_xml, label_encoding = "UTF-8")
}
\arguments{
\item{parsed_xml}{an object of class XMLInternalDocument}

\item{label_encoding}{}
}
\value{
dataframe
}
\description{
Title
}
\examples{

}
