% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_file_info.R
\name{ox_file_info}
\alias{ox_file_info}
\title{File info in a list}
\usage{
ox_file_info(parsed_xml)
}
\arguments{
\item{parsed_xml}{An object of class \code{XMLInternalDocument}, as returned
by \code{XML::xmlParse()}.}
}
\value{
A list.
}
\description{
Returns a list with information on the .xml file, from a parsed OpenClinica
odm1.3 .xml export file.
}
\examples{
# The example odm1.3 xml file address
my_file <- system.file("extdata",
                       "odm1.3_clinical_ext_example.xml",
                       package = "ox",
                       mustWork = TRUE)

# Parsing the xml file
library(XML)
doc <- xmlParse(my_file)

# File info in a list
file_info <- ox_file_info(doc)
file_info
}
