% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_form_def.R
\name{ox_form_def}
\alias{ox_form_def}
\title{Form definitions in a dataframe}
\usage{
ox_form_def(parsed_xml)
}
\arguments{
\item{parsed_xml}{An object of class \code{XMLInternalDocument}, as returned
by \code{XML::xmlParse()}.}
}
\value{
A dataframe.
}
\description{
Returns a dataframe with study form definitions from a parsed OpenClinica
odm1.3 .xml export file.
}
\examples{
# The example odm1.3 xml file address
my_file <- system.file("extdata",
                       "odm1.3_clinical_ext_example.xml",
                       package = "ox",
                       mustWork = TRUE)

# Parsing the xml file
library(XML)
doc <- xmlParse(my_file)

# Form definitions in a dataframe
form_def <- ox_form_def(doc)
View(event_def)
}
