% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_form_ref.R
\name{ox_form_ref}
\alias{ox_form_ref}
\title{Title Form references}
\usage{
ox_form_ref(parsed_xml)
}
\arguments{
\item{parsed_xml}{an object of class XMLInternalDocument}
}
\value{
dataframe
}
\description{
Title Form references
}
\examples{

}
