% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_global_vars.R
\name{ox_global_vars}
\alias{ox_global_vars}
\title{Title Global study variables}
\usage{
ox_global_vars(parsed_xml)
}
\arguments{
\item{parsed_xml}{an object of class XMLInternalDocument}
}
\value{
dataframe
}
\description{
Title Global study variables
}
\examples{

}
