% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_group_def.R
\name{ox_group_def}
\alias{ox_group_def}
\title{Title ItemGroup definitions}
\usage{
ox_group_def(parsed_xml)
}
\arguments{
\item{parsed_xml}{an object of class XMLInternalDocument}
}
\value{
dataframe
}
\description{
Title ItemGroup definitions
}
\examples{

}
