% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_group_ref.R
\name{ox_group_ref}
\alias{ox_group_ref}
\title{Title ItemGroup references}
\usage{
ox_group_ref(parsed_xml)
}
\arguments{
\item{parsed_xml}{an object of class XMLInternalDocument}
}
\value{
dataframe
}
\description{
Title ItemGroup references
}
\examples{

}
