% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_group_repeat.R
\name{ox_group_repeat}
\alias{ox_group_repeat}
\title{Title}
\usage{
ox_group_repeat(parsed_xml)
}
\arguments{
\item{parsed_xml}{an object of class XMLInternalDocument}
}
\value{
dataframe
}
\description{
Title
}
\examples{

}
