% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_group_to_df.R
\name{ox_group_to_df}
\alias{ox_group_to_df}
\title{Title Creates a dataframe with data from an ItemGroup}
\usage{
ox_group_to_df(item_data, GroupOID)
}
\arguments{
\item{GroupOID}{ItemGroupOID as character.}
}
\value{
dataframe
}
\description{
Title Creates a dataframe with data from an ItemGroup
}
\examples{

}
