% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_item_data.R
\name{ox_item_data}
\alias{ox_item_data}
\title{Title Clinical data}
\usage{
ox_item_data(parsed_xml, value_encoding = "UTF-8")
}
\arguments{
\item{parsed_xml}{an object of class XMLInternalDocument}

\item{value_encoding}{}
}
\value{
dataframe
}
\description{
Title Clinical data
}
\examples{

}
