% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_item_def.R
\name{ox_item_def}
\alias{ox_item_def}
\title{Title Item definitions}
\usage{
ox_item_def(parsed_xml, comment_encoding = "UTF-8")
}
\arguments{
\item{parsed_xml}{an object of class XMLInternalDocument}

\item{comment_encoding}{}
}
\value{
dataframe
}
\description{
Title Item definitions
}
\examples{

}
