% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_item_ref.R
\name{ox_item_ref}
\alias{ox_item_ref}
\title{Title Item references}
\usage{
ox_item_ref(parsed_xml)
}
\arguments{
\item{parsed_xml}{an object of class XMLInternalDocument}
}
\value{
dataframe
}
\description{
Title Item references
}
\examples{

}
