% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_measurement_units.R
\name{ox_measurement_units}
\alias{ox_measurement_units}
\title{Title Measurement units}
\usage{
ox_measurement_units(parsed_xml, name_encoding = "UTF-8")
}
\arguments{
\item{parsed_xml}{an object of class XMLInternalDocument}

\item{name_encoding}{}
}
\value{
dataframe
}
\description{
Title Measurement units
}
\examples{

}
