% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_metadata.R
\name{ox_metadata}
\alias{ox_metadata}
\title{Metadata in a list of dataframes}
\usage{
ox_metadata(parsed_xml)
}
\arguments{
\item{parsed_xml}{An object of class \code{XMLInternalDocument}, as returned
by \code{XML::xmlParse()}.}
}
\value{
A list containing the following elements:

\itemize{
  \item \code{file_info}: a list with .xml file information.
  \item \code{global_vars}: a list with global study variables.
  \item \code{event_def}: a dataframe of event definitions.
  \item \code{event_ref}: a dataframe of event references.
  \item \code{form_def}: a dataframe of form definitions.
  \item \code{form_ref}: a dataframe of form references.
  \item \code{group_def}: a dataframe of (item) group definitions.
  \item \code{group_ref}: a dataframe of (item) group references.
  \item \code{codelist}: a dataframe of codelists.
  \item \code{codelist_item}: a dataframe of codelist items.
  \item \code{codelist_ref}: a dataframe of codelist references.

  \item \code{units}: a dataframe of unit definitions.
  \item \code{sites}: a dataframe of study sites.
  \item \code{subjects}: a dataframe of study subjects.
}
}
\description{
Returns a list of dataframes containing metadata, from a parsed OpenClinica
odm1.3 .xml export file provided as argument.
}
\examples{
# The example odm1.3 xml file address
my_file <- system.file("extdata",
                       "odm1.3_full_example.xml",
                       package = "ox",
                       mustWork = TRUE)

# Parsing the xml file
library(XML)
doc <- xmlParse(my_file)

# Get metadata
md <- ox_metadata(doc)
class(md)
names(md)

# Accessing metadata elements:
# Event definitions
head(md$event_def)

# Codelist items
head(md$codelist_item)

# etc.
}
