% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_sites.R
\name{ox_sites}
\alias{ox_sites}
\title{Sites in a dataframe}
\usage{
ox_sites(parsed_xml)
}
\arguments{
\item{parsed_xml}{An object of class \code{XMLInternalDocument}, as returned
by \code{XML::xmlParse()}.}
}
\value{
A dataframe.
}
\description{
Returns a dataframe with study sites from a parsed OpenClinica
odm1.3 .xml export file.
}
\examples{
# The example odm1.3 xml file address
my_file <- system.file("extdata",
                       "odm1.3_full_example.xml",
                       package = "ox",
                       mustWork = TRUE)

# Parsing the xml file
library(XML)
doc <- xmlParse(my_file)

# Sites in a dataframe
sites <- ox_sites(doc)
head(sites)

}
