% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_subject_data.R
\name{ox_subject_data}
\alias{ox_subject_data}
\title{Title Subject data}
\usage{
ox_subject_data(parsed_xml)
}
\arguments{
\item{parsed_xml}{an object of class XMLInternalDocument}
}
\value{
dataframe
}
\description{
Title Subject data
}
\examples{

}
