% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_xtract_group.R
\name{ox_xtract_group}
\alias{ox_xtract_group}
\title{Title Creates a dataframe with data from an ItemGroup}
\usage{
ox_xtract_group(ox_obj, group, define_factors = FALSE,
  use_item_names = FALSE)
}
\arguments{
\item{group}{ItemGroupOID as character.}
}
\value{
dataframe
}
\description{
Title Creates a dataframe with data from an ItemGroup
}
\examples{

}
