% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ox.R
\name{summary.ox}
\alias{summary.ox}
\title{Summary for \code{ox} objects}
\usage{
\method{summary}{ox}(ox_obj)
}
\arguments{
\item{ox_obj}{An object of class \code{ox}, as returned by \code{ox()}.}
}
\value{
A list with the following components:

\itemize{
  \item \code{numbers}: numbers of datapoints, subjects, sites, events, forms,
groups, and items, in  \code{ox_obj$data}.
  \item \code{events}: unique values of \code{event_oid} in \code{ox_obj$data}.
  \item \code{forms}: unique values of \code{form_oid} in \code{ox_obj$data}.
  \item \code{groups}: unique values of \code{group_oid} in \code{ox_obj$data}.
  \item\code{assessments}: table of \code{event_oid} by \code{group_oid}, in
  \code{ox_obj$data}, Shows the number of datapoints (rows in
\code{ox_obj$data}) per event-group combination.
}
}
\description{
Produces a summary of an \code{ox} object, including minimal information on
datapoints, subjects, items, groups, forms and events.
}
\examples{
# The example odm1.3 xml file address
my_file <- system.file("extdata",
                       "odm1.3_clinical_ext_example.xml",
                       package = "ox",
                       mustWork = TRUE)

# Parsing the xml file
library(XML)
doc <- xmlParse(my_file)

# Creating ox object and getting summary info
my_ox_obj <- ox(doc)
summary(my_ox_obj)

}
