% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_fun_args}
\alias{validate_fun_args}
\title{Get index of an \code{.else} argument}
\usage{
validate_fun_args(.f, args)
}
\arguments{
\item{.f}{(\code{function}) evaluated as \code{.f(...)}. Must return single logical value}

\item{args}{\code{\link{list}}\cr arguments}
}
\value{
\code{NULL}
}
\description{
Get index of an \code{.else} argument. Returns index of the potential
}
