% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which_else}
\alias{which_else}
\title{Get index of an \code{.else} argument}
\usage{
which_else(args)
}
\arguments{
\item{args}{\code{\link{list}}\cr arguments}
}
\value{
\code{integer(1)} or \code{integer(0)} when no unnamed list
elements.
}
\description{
Get index of an \code{.else} argument. Returns index of the potential
}
