% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie.R
\name{ie}
\alias{ie}
\alias{nie}
\title{Short hand if-else}
\usage{
ie(fun, ..., .else)

nie(fun, ..., .else)
}
\arguments{
\item{fun}{\code{\link{function}}\cr}

\item{...}{\cr
arguments passed to the \code{fun} in the same order.}

\item{.else}{\cr
negative-replacement. NOTE that if \code{.else} is not specified directly by naming
argument then the last unnamed argument (from \code{...}) will be considered as a
replacement (it will not be used in the \code{fun}).}
}
\description{
Short hand if-else functions for easy values switching.
\code{ie} evaluates function returning a single logical value and depending on the
result returning:
\itemize{
\item on \code{TRUE} - \code{ie} returns \code{...[1]}. First element in \code{...} is considered
as a  positive-replacement.
\item on \code{FALSE} - \code{ie} returns \code{.else} (negative-replacement).
}
}
\details{
To invert switch one can use \code{nie} which is equivalent of
\code{ie(Negate(fun), ..., .else)}.
}
\examples{
# return don't switch if TRUE
ie(is.null, NULL, "y")
ie(grepl, x = "some text", pattern = "word", .else = "does not contain a word")

# return replacement `y` if is TRUE
nie(is.null, NULL, "y")
nie(grepl, x = "some text", pattern = "word", .else = "does not contain a word")
}
