% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie.R
\name{ie}
\alias{ie}
\alias{nie}
\title{Short hand \code{ifelse}}
\usage{
ie(fun, ..., .else)

nie(fun, ..., .else)
}
\arguments{
\item{fun}{\code{\link{function}}\cr}

\item{...}{\cr
arguments passed to the \code{fun} in the same order.}

\item{.else}{\cr
replacement object. NOTE that if \code{.else} is not specified as named argument
else the last unnamed argument will be taken as a replacement (it will
not be used in the \code{fun}).}
}
\description{
Short hand \code{ifelse} functions for easy values switching.
\itemize{
\item \code{ie} (if-else) - returns first argument from \code{...} if \code{TRUE},
returns \code{.else} otherwise.
\item \code{nie} (negative it-else) - returns \code{.else} if \code{TRUE} and first argument
from \code{...} otherwise.
}
}
\examples{
# return don't switch if TRUE
ie(is.null, NULL, "y")

ie(function(x, min, max) nchar(x) >= min && nchar(x) <= max,
   x = "x", min = 0, max = 1,
   .else = "test passed")

# return replacement `y` if is TRUE
nie(is.null, NULL, "y")

nie(function(x, min, max) nchar(x) >= min && nchar(x) <= max,
   x = "test", min = 20, max = Inf,
   .else = "replacement")
}
