% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox.R
\name{ox}
\alias{ox}
\title{Creates an \code{ox} object}
\usage{
ox(parsed_xml)
}
\arguments{
\item{parsed_xml}{An object of class \code{XMLInternalDocument}, as returned
by \code{XML::xmlParse()}.}
}
\value{
An object of class \code{ox}, which is a list of two elements:

\itemize{
  \item \code{data}: a dataframe containig all clinical data, as returned by
  \code{ox_item_data()}.
  \item \code{metadata}: a list, as returned by \code{ox_metadata()}. See
  \code{\link{ox_metadata}} for more details.
}
}
\description{
Returns an object of class \code{ox}, containing data and metadata, from a
parsed OpenClinica odm1.3 .xml export file provided as argument.
}
\details{
A superficial view of the contents of an \code{ox} object can be
obtained with \code{\link{ox_info}}.
}
\examples{
# The example odm1.3 xml file address
my_file <- system.file("extdata",
                       "odm1.3_clinical_ext_example.xml",
                       package = "ox",
                       mustWork = TRUE)

# Parsing the xml file
library(XML)
doc <- xmlParse(my_file)

# Create ox object
my_study <- ox(doc)
class(my_study)
names(my_study)

# The data element
head(my_study$data)

# Elements (names) in metadata
names(my_study$metadata)

# Accessing metadata elements:
# Event definitions
head(my_study$metadata$event_def)

# Codelist items
head(my_study$metadata$codelist_item)

# etc.
}
