% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_codelist_item.R
\name{ox_codelist_item}
\alias{ox_codelist_item}
\title{Codelists items in a dataframe}
\usage{
ox_codelist_item(parsed_xml)
}
\arguments{
\item{parsed_xml}{An object of class \code{XMLInternalDocument}, as returned
by \code{XML::xmlParse()}.}
}
\value{
A dataframe.
}
\description{
Returns a dataframe with study codelist items from a parsed OpenClinica
odm1.3_full .xml export file.
}
\examples{
# The example odm1.3 xml file address
my_file <- system.file("extdata",
                       "odm1.3_full_example.xml",
                       package = "ox",
                       mustWork = TRUE)

# Parsing the xml file
library(XML)
doc <- xmlParse(my_file)

# Codelists items in a dataframe
codelist_items <- ox_codelist_item(doc)
head(codelist_items)
}
