% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_codelist_ref.R
\name{ox_codelist_ref}
\alias{ox_codelist_ref}
\title{Codelist references in a dataframe}
\usage{
ox_codelist_ref(parsed_xml)
}
\arguments{
\item{parsed_xml}{An object of class \code{XMLInternalDocument}, as returned
by \code{XML::xmlParse()}.}
}
\value{
A dataframe.
}
\description{
Returns a dataframe with study codelist references from a parsed OpenClinica
odm1.3 .xml export file.
}
\examples{
# The example odm1.3 xml file address
my_file <- system.file("extdata",
                       "odm1.3_clinical_ext_example.xml",
                       package = "ox",
                       mustWork = TRUE)

# Parsing the xml file
library(XML)
doc <- xmlParse(my_file)

# Form references in a dataframe
codelist_ref <- ox_codelist_ref(doc)
View(codelist_ref)

}
