% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_event_def.R
\name{ox_event_def}
\alias{ox_event_def}
\title{Title}
\usage{
ox_event_def(parsed_xml, name_encoding = "UTF-8")
}
\arguments{
\item{parsed_xml}{an object of class XMLInternalDocument}

\item{name_encoding}{}
}
\value{
dataframe
}
\description{
Title
}
\examples{

}
