% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_event_def.R
\name{ox_event_def}
\alias{ox_event_def}
\title{Event definitions in a dataframe}
\usage{
ox_event_def(parsed_xml)
}
\arguments{
\item{parsed_xml}{An object of class \code{XMLInternalDocument}, as returned
by \code{XML::xmlParse()}.}
}
\value{
A dataframe.
}
\description{
Returns a dataframe with study event definitions from a parsed OpenClinicaC
odm1.3 .xml export file.
}
\examples{

# Get the example file address
file_address <- system.file("extdata",
                            "odm1.3_clinical_ext_example.xml",
                            package = "ox",
                            mustWork = TRUE)

# Parsing the xml file
library(XML)
doc <- xmlParse(file_address)

# Event definitions in a dataframe
ox_event_def(doc)

}
