% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_event_ref.R
\name{ox_event_ref}
\alias{ox_event_ref}
\title{Title}
\usage{
ox_event_ref(parsed_xml, simplify = FALSE)
}
\arguments{
\item{parsed_xml}{an object of class XMLInternalDocument}

\item{simplify}{logical}
}
\value{
dataframe
}
\description{
Title
}
\examples{

}
