% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_form_def.R
\name{ox_form_def}
\alias{ox_form_def}
\title{Title Form definitions}
\usage{
ox_form_def(parsed_xml)
}
\arguments{
\item{parsed_xml}{an object of class XMLInternalDocument}
}
\value{
dataframe
}
\description{
Title Form definitions
}
\examples{

}
