% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_group_def.R
\name{ox_group_def}
\alias{ox_group_def}
\title{Group definitions in a dataframe}
\usage{
ox_group_def(parsed_xml)
}
\arguments{
\item{parsed_xml}{An object of class \code{XMLInternalDocument}, as returned
by \code{XML::xmlParse()}.}
}
\value{
A dataframe.
}
\description{
Returns a dataframe with study item group definitions from a parsed OpenClinica
odm1.3 .xml export file.
}
\examples{
# The example xml file address
my_file <- system.file("extdata",
                       "odm1.3_full_example.xml",
                       package = "ox",
                       mustWork = TRUE)

# Parsing the xml file
library(XML)
doc <- xmlParse(my_file)

# Event definitions in a dataframe
group_def <- ox_group_def(doc)
head(group_def)
}
