% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_group_ref.R
\name{ox_group_ref}
\alias{ox_group_ref}
\title{Group references in a dataframe}
\usage{
ox_group_ref(parsed_xml)
}
\arguments{
\item{parsed_xml}{An object of class \code{XMLInternalDocument}, as returned
by \code{XML::xmlParse()}.}
}
\value{
A dataframe.
}
\description{
Returns a dataframe with study (item) group references from a parsed OpenClinica
odm1.3 .xml export file.
}
\examples{
# The example odm1.3 xml file address
my_file <- system.file("extdata",
                       "odm1.3_clinical_ext_example.xml",
                       package = "ox",
                       mustWork = TRUE)

# Parsing the xml file
library(XML)
doc <- xmlParse(my_file)

# Form references in a dataframe
group_ref <- ox_group_ref(doc)
head(group_ref)

}
