% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_group_repeat.R
\name{ox_group_repeat}
\alias{ox_group_repeat}
\title{Group repetitions in a dataframe}
\usage{
ox_group_repeat(parsed_xml)
}
\arguments{
\item{parsed_xml}{An object of class \code{XMLInternalDocument}, as returned
by \code{XML::xmlParse()}.}
}
\value{
A dataframe.
}
\description{
Returns a dataframe with study item group repetitions from a parsed OpenClinica
odm1.3 .xml export file.
}
\examples{
# The example xml file address
my_file <- system.file("extdata",
                       "odm1.3_full_example.xml",
                       package = "ox",
                       mustWork = TRUE)

# Parsing the xml file
library(XML)
doc <- xmlParse(my_file)

# Group repetitions in a dataframe
group_repeats <- ox_group_repeat(doc)
head(group_repeats)

}
