% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_item_def.R
\name{ox_item_def}
\alias{ox_item_def}
\title{Item definitions in a dataframe}
\usage{
ox_item_def(parsed_xml)
}
\arguments{
\item{parsed_xml}{An object of class \code{XMLInternalDocument}, as returned
by \code{XML::xmlParse()}.}
}
\value{
A dataframe.
}
\description{
Returns a dataframe with study item definitions from a parsed OpenClinica
odm1.3 .xml export file.
}
\examples{
# The example xml file address
file <- system.file("extdata",
                    "odm1.3_clinical_ext_example.xml",
                    package = "ox",
                    mustWork = TRUE)

# Parsing the xml file
library(XML)
doc <- xmlParse(file_address)

# Item definitions in a dataframe
item_def <- ox_item_def(doc)
View(item_def)
}
