% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_measurement_units.R
\name{ox_measurement_units}
\alias{ox_measurement_units}
\title{Measurement units in a dataframe}
\usage{
ox_measurement_units(parsed_xml)
}
\arguments{
\item{parsed_xml}{An object of class \code{XMLInternalDocument}, as returned
by \code{XML::xmlParse()}.}
}
\value{
A dataframe.
}
\description{
Returns a dataframe with measurement units from a parsed OpenClinica
odm1.3 .xml export file.
}
\examples{
# The example odm1.3 xml file address
my_file <- system.file("extdata",
                       "odm1.3_full_example.xml",
                       package = "ox",
                       mustWork = TRUE)

# Parsing the xml file
library(XML)
doc <- xmlParse(my_file)

# Measurement units in a dataframe
measurement_units <- ox_measurement_units(doc)
head(measurement_units)
}
