% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_metadata.R
\name{ox_metadata}
\alias{ox_metadata}
\title{Metadata in a list of dataframes}
\usage{
ox_metadata(parsed_xml)
}
\arguments{
\item{parsed_xml}{An object of class \code{XMLInternalDocument}, as returned
by \code{XML::xmlParse()}.}
}
\value{
A list of elements, most of which are dataframes, collectively
describing the study data structure (events, forms, item groups,
items, codelists, measurement units, sites, and subjects).
}
\description{
Returns a list of dataframes containing metadata, from a parsed OpenClinica
odm1.3 .xml export file provided as argument.
}
\examples{
# The example odm1.3 xml file address
my_file <- system.file("extdata",
                       "odm1.3_clinical_ext_example.xml",
                       package = "ox",
                       mustWork = TRUE)

# Parsing the xml file
library(XML)
doc <- xmlParse(my_file)

# Get metadata
md <- ox_metadata(doc)
class(md)
names(md)

# Accessing metadata elements:
# Event definitions
View(md$event_def)

# Codelist items
View(md$codelist_item)

# etc.
}
