% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_subject_data.R
\name{ox_subject_data}
\alias{ox_subject_data}
\title{Subject data in a dataframe}
\usage{
ox_subject_data(parsed_xml)
}
\arguments{
\item{parsed_xml}{An object of class \code{XMLInternalDocument}, as returned
by \code{XML::xmlParse()}.}
}
\value{
A dataframe.
}
\description{
Returns a dataframe with study subjects data from a parsed OpenClinica
odm1.3 .xml export file.
}
\examples{
# The example xml file address
file <- system.file("extdata",
                    "odm1.3_clinical_ext_example.xml",
                    package = "ox",
                    mustWork = TRUE)

# Parsing the xml file
library(XML)
doc <- xmlParse(file_address)

# Subject data in a dataframe
subjects <- ox_subject_data(doc)
View(subjects)
}
